requirements.txt
aerich==0.7.2
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.2.0
asyncpg==0.29.0
bcrypt==4.0.1
black==23.12.1
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
dictdiffer==0.9.0
django-environ==0.11.2
dnspython==2.4.2
ecdsa==0.18.0
fastapi==0.109.0
gunicorn==21.2.0
h11==0.14.0
idna==3.6
iso8601==1.1.0
isort==5.13.2
mypy-extensions==1.0.0
packaging==23.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.1.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pymongo==4.6.1
pypika-tortoise==0.1.6
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
redis==5.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.35.1
tomlkit==0.12.3
tortoise-orm==0.20.0
typing_extensions==4.9.0
uvicorn==0.25.0
